requirements.txt
absl-py==0.7.1
appdirs==1.4.3
appnope==0.1.0
asciimatics==1.11.0
astor==0.8.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
configobj==5.0.6
configparser==3.8.1
contextlib2==0.5.5
cycler==0.10.0
ddt==1.2.1
decorator==4.4.0
defusedxml==0.6.0
distro==1.4.0
dvc==0.54.1
entrypoints==0.3
Flask==1.1.1
funcy==1.13
future==0.17.1
gast==0.2.2
gevent==1.4.0
gitdb==0.6.4
gitdb2==2.0.5
GitPython==3.0.0
google-pasta==0.1.7
grandalf==0.6
greenlet==0.4.15
grpcio==1.22.0
h5py==2.9.0
humanize==0.5.1
idna==2.8
imagecodecs==2018.10.30
importlib-metadata==0.19
imutils==0.5.1
inflect==2.1.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonpath-ng==1.4.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
locustio==0.11.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
msgpack==0.6.1
nanotime==0.5.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.0
numpy==1.17.0
opencv-contrib-python==3.4.3.18
opencv-python==3.4.3
opt-einsum==3.1.0
packaging==19.1
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.5.9
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plac==1.1.0
pluggy==0.12.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.6
pycparser==2.19
pyfiglet==0.8.post1
Pygments==2.4.2
pylint==2.4.3
pyparsing==2.4.2
pypiwin32==223
PyQt5==5.10
pyrsistent==0.15.4
pytest==5.0.1
python-dateutil==2.8.0
python-mnist==0.6
pytz==2019.2
pywin32==224
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.2
requests==2.22.0
ruamel.yaml==0.15.100
schema==0.7.0
scipy==1.3.1
Send2Trash==1.5.0
shortuuid==0.5.0
sip==4.19.8
six==1.12.0
smmap==0.9.0
smmap2==2.0.5
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.33.0
traitlets==4.3.2
treelib==1.5.5
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
win-unicode-console==0.5
wrapt==1.11.2
zc.lockfile==2.0
zipp==0.5.2
